adds save button to question Id input and refactored survey saving/publishing logic
